Move icon/surface/ to surface/icon/
Update absolute imports in test files, icon_physics.py, and parameters.py.
