+++ jcm/physics/icon/icon_physics.py
@@ -25,8 +25,8 @@ from jcm.physics.icon.icon_physics_data import RadiationData
 from jcm.physics.convection.icon import tiedtke_nordeng_convection
 from jcm.physics.clouds.icon import shallow_cloud_scheme, cloud_microphysics
 from jcm.physics.icon.parameters import Parameters
-from jcm.physics.icon.surface import surface_physics_step, initialize_surface_state
-from jcm.physics.icon.surface.surface_types import AtmosphericForcing
+from jcm.physics.surface.icon import surface_physics_step, initialize_surface_state
+from jcm.physics.surface.icon.surface_types import AtmosphericForcing
 from jcm.physics.gravity_waves.icon import gravity_wave_drag
 from jcm.physics.chemistry.icon import simple_chemistry
 from jcm.physics.aerosol.icon.simple_aerosol import get_simple_aerosol
